refactor: TicketFactory(Fixed,Random)을 Factory 패키지로 이동
