+++ src/main/java/lotto/controller/LottoController.java
@@ -2,7 +2,7 @@ package lotto.controller;
 
 import java.util.List;
 import java.util.Scanner;
-import lotto.domain.FixedTicketFactory;
+import lotto.domain.ticketFactory.FixedTicketFactory;
 import lotto.domain.LottoNumber;
 import lotto.domain.LottoTicket;
 import lotto.domain.LottoTickets;
