+++ src/test/java/lotto/domain/ticketFactory/RandomTicketFactoryTest.java
@@ -1,8 +1,8 @@
-package lotto.domain;
+package lotto.domain.ticketFactory;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.Set;
+import lotto.domain.LottoTicket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,7 +11,7 @@ public class RandomTicketFactoryTest {
     @DisplayName("티켓 팩토리에서 생성한 티켓 넘버의 개수가 6개인지 확인")
     @Test
     void checkSizeOfTicketNumbers() {
-        Set<LottoNumber> lottoNumbers = RandomTicketFactory.shuffle();
-        assertThat(lottoNumbers.size()).isEqualTo(6);
+        LottoTicket lottoTicket = RandomTicketFactory.makeTicket();
+        assertThat(lottoTicket.getLottoNumbers().size()).isEqualTo(6);
     }
 }
